refactor(core): unify HTTP routing on OpenAPI app, drop dead route factories
createSnaApp() already delegates to createOpenApiApp(), so the legacy
routes/agent.ts createAgentRoutes (888 lines) and routes/chat.ts
createChatRoutes (172 lines) factories were exported but never mounted
— dead code that duplicated every route the OpenAPI app already serves.

- Extract `runOnce` (the only useful symbol in agent.ts) to its own
  module server/run-once.ts and update ws.ts + openapi.ts to import
  from there.
- Drop `configDir: opts.cwd` from the pool path in runOnce; it pointed
  CODEX_HOME / CLAUDE_CONFIG_DIR at the working tree, which has no
  auth.json, so pooled Codex run-once was effectively broken.
- Gut routes/agent.ts and routes/chat.ts to deprecated re-export shims
  so the existing subpath exports keep resolving for any consumer that
  imported `runOnce` from `@sna-sdk/core/server/routes/agent`.
- Drop the dead `createAgentRoutes` / `createChatRoutes` exports from
  server/index.ts; export the relocated `runOnce` and its types instead.

OpenAPI app cleanups so the spec served at /openapi.json matches reality:

- Resolve `info.version` from packages/core/package.json at module load
  via createRequire (was hardcoded to a stale "0.11.3"; now reflects
  whatever ships).
- Promote `POST /agent/list-models` from a plain `app.post()` to a
  full `app.openapi()` route so it shows up in the spec and Swagger UI.
- Add the SSE route `GET /agent/events` to the OpenAPI app — the SSE
  stream only lived in the dead routes/agent.ts before, so the live
  server never actually exposed it. HTTP-only clients can now subscribe
  to a session's events without a WebSocket dependency.
- Replace the inline run-once handler with a single delegate to
  `runOnce()`, so HTTP and WS share one code path (this is what
  picks up the configDir fix above).

Tests: api-parity now reads from routes/openapi.ts (the live router)
instead of the gutted agent.ts/chat.ts, and includes the
`sessions.update`, `agent.completion`, `agent.list-models` ops it was
previously missing.

core/README.md exports table: drop the now-empty `routes/agent` and
`routes/chat` subpath entries (the shims still exist for backwards
compat but have nothing to import), bring the `providers` line in line
with the actual export list.
