+++ packages/core/src/server/routes/agent.ts
@@ -1,888 +1,11 @@
 /**
- * Agent routes — HTTP API for spawning and communicating with agent providers.
+ * @deprecated The HTTP agent routes now live in `routes/openapi.ts` and are
+ * mounted automatically by `createSnaApp()`. This module survives only to
+ * keep the `@sna-sdk/core/server/routes/agent` subpath export working for
+ * consumers who imported `runOnce` from here historically.
  *
- * Routes:
- *   POST /start?session=<id>   — start agent in a session (default: "default")
- *   POST /send?session=<id>    — send a message to a session
- *   GET  /events?session=<id>  — SSE stream for a session
- *   POST /kill?session=<id>    — kill agent in a session
- *   GET  /status?session=<id>  — check session status
- *
- *   POST /sessions             — create a new session
- *   GET  /sessions             — list all sessions
- *   DELETE /sessions/:id       — remove a session
- *
- *   POST /run-once             — one-shot: spawn → execute → return result → cleanup
+ * Import `runOnce` from `@sna-sdk/core/server` directly instead.
  */
 
-import { Hono } from "hono";
-import { streamSSE } from "hono/streaming";
-import {
-  getProvider,
-  spawnWithPool,
-  type AgentEvent,
-  type AgentProcess,
-} from "../../core/providers/index.js";
-import type { SessionConfig } from "../session-manager.js";
-import { logger } from "../../lib/logger.js";
-import { getDb } from "../../db/schema.js";
-import { SessionManager } from "../session-manager.js";
-import { buildCanonicalFromDb } from "../../history/canonical.js";
-import { httpJson } from "../api-types.js";
-import { saveEmbeds } from "../image-store.js";
-import { insertChatMessage } from "../../db/chat-messages.js";
-import { formatEmbedRef } from "../../history/embed-refs.js";
-import type { EmbedRecord } from "../../history/types.js";
-import { getConfig } from "../../config.js";
-import { completion, type CompletionOptions } from "../../core/completion.js";
-
-/** Helper: read session ID from query string, default "default". */
-function getSessionId(c: { req: { query: (k: string) => string | undefined } }): string {
-  return c.req.query("session") ?? "default";
-}
-
-// ── run-once shared logic ─────────────────────────────────────────
-
-export interface RunOnceOptions {
-  message: string;
-  model?: string;
-  systemPrompt?: string;
-  appendSystemPrompt?: string;
-  permissionMode?: string;
-  cwd?: string;
-  timeout?: number;
-  provider?: string;
-  extraArgs?: string[];
-  env?: Record<string, string>;
-  /** Provider-specific options (e.g. `{ omlxBaseUrl: "http://..." }`). */
-  providerOptions?: Record<string, unknown>;
-}
-
-export interface RunOnceResult {
-  result: string;
-  usage: Record<string, unknown> | null;
-}
-
-/**
- * One-shot agent execution: create temp session → spawn → wait for result → cleanup.
- * Used by both HTTP POST /run-once and WS agent.run-once.
- */
-export async function runOnce(
-  sessionManager: SessionManager,
-  opts: RunOnceOptions,
-): Promise<RunOnceResult> {
-  const sessionId = `run-once-${crypto.randomUUID().slice(0, 8)}`;
-  const timeout = opts.timeout ?? getConfig().runOnceTimeoutMs;
-
-  const session = sessionManager.createSession({
-    id: sessionId,
-    label: "run-once",
-    cwd: opts.cwd ?? process.cwd(),
-  });
-
-  const cfg = getConfig();
-  const provider = getProvider(opts.provider ?? cfg.defaultProvider);
-
-  const extraArgs: string[] = opts.extraArgs ? [...opts.extraArgs] : [];
-  if (opts.systemPrompt) extraArgs.push("--system-prompt", opts.systemPrompt);
-  if (opts.appendSystemPrompt) extraArgs.push("--append-system-prompt", opts.appendSystemPrompt);
-
-  const proc = await spawnWithPool(provider, {
-    cwd: session.cwd,
-    prompt: opts.message,
-    model: opts.model ?? cfg.model,
-    permissionMode: (opts.permissionMode as any) ?? cfg.defaultPermissionMode,
-    configDir: opts.cwd,
-    env: { ...opts.env, SNA_SESSION_ID: sessionId },
-    extraArgs,
-    providerOptions: opts.providerOptions,
-    systemPrompt: opts.systemPrompt,
-    appendSystemPrompt: opts.appendSystemPrompt,
-  });
-
-  sessionManager.setProcess(sessionId, proc);
-
-  try {
-    const result = await new Promise<RunOnceResult>((resolve, reject) => {
-      const texts: string[] = [];
-      let usage: Record<string, unknown> | null = null;
-
-      const timer = setTimeout(() => {
-        reject(new Error(`run-once timed out after ${timeout}ms`));
-      }, timeout);
-
-      const unsub = sessionManager.onSessionEvent(sessionId, (_cursor, e) => {
-        if (e.type === "assistant" && e.message) {
-          texts.push(e.message);
-        }
-        if (e.type === "complete") {
-          clearTimeout(timer);
-          unsub();
-          usage = (e.data as Record<string, unknown>) ?? null;
-          resolve({ result: texts.join("\n"), usage });
-        }
-        if (e.type === "error") {
-          clearTimeout(timer);
-          unsub();
-          reject(new Error(e.message ?? "Agent error"));
-        }
-      });
-    });
-
-    return result;
-  } finally {
-    sessionManager.killSession(sessionId);
-    sessionManager.removeSession(sessionId);
-  }
-}
-
-export function createAgentRoutes(sessionManager: SessionManager) {
-  const app = new Hono();
-
-  // ── Session CRUD ──────────────────────────────────────────────
-
-  // POST /sessions — create a new session
-  app.post("/sessions", async (c) => {
-    const body = (await c.req.json().catch(() => ({}))) as {
-      id?: string;
-      label?: string;
-      cwd?: string;
-      meta?: Record<string, unknown>;
-    };
-
-    try {
-      const session = sessionManager.createSession({
-        id: body.id,
-        label: body.label,
-        cwd: body.cwd,
-        meta: body.meta,
-      });
-
-      logger.log("route", `POST /sessions → created "${session.id}"`);
-      return httpJson(c, "sessions.create", { status: "created", sessionId: session.id, label: session.label, meta: session.meta });
-    } catch (e: any) {
-      logger.err("err", `POST /sessions → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 409);
-    }
-  });
-
-  // GET /sessions — list all sessions. Pass `?include=chain` to embed each
-  // session's full RuntimeSession audit chain.
-  app.get("/sessions", (c) => {
-    const include = c.req.query("include");
-    return httpJson(c, "sessions.list", {
-      sessions: sessionManager.listSessions({ includeRuntimeChain: include === "chain" }),
-    });
-  });
-
-  // DELETE /sessions/:id — remove a session
-  app.delete("/sessions/:id", (c) => {
-    const id = c.req.param("id");
-    if (id === "default") {
-      return c.json({ status: "error", message: "Cannot remove default session" }, 400);
-    }
-    const removed = sessionManager.removeSession(id);
-    if (!removed) {
-      return c.json({ status: "error", message: "Session not found" }, 404);
-    }
-    logger.log("route", `DELETE /sessions/${id} → removed`);
-    return httpJson(c, "sessions.remove", { status: "removed" });
-  });
-
-  // PATCH /sessions/:id — update session metadata (label, meta, cwd)
-  app.patch("/sessions/:id", async (c) => {
-    const id = c.req.param("id");
-    const body = (await c.req.json().catch(() => ({}))) as {
-      label?: string;
-      meta?: Record<string, unknown>;
-      cwd?: string;
-    };
-    try {
-      sessionManager.updateSession(id, {
-        label: body.label,
-        meta: body.meta,
-        cwd: body.cwd,
-      });
-      logger.log("route", `PATCH /sessions/${id} → updated`);
-      return httpJson(c, "sessions.update", { status: "updated", session: id });
-    } catch (e: any) {
-      logger.err("err", `PATCH /sessions/${id} → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 404);
-    }
-  });
-
-  // ── One-shot execution ─────────────────────────────────────────
-
-  // POST /run-once — spawn → execute → return result → cleanup
-  app.post("/run-once", async (c) => {
-    const body = (await c.req.json().catch(() => ({}))) as RunOnceOptions;
-    if (!body.message) {
-      return c.json({ status: "error", message: "message is required" }, 400);
-    }
-    try {
-      const result = await runOnce(sessionManager, body);
-      return httpJson(c, "agent.run-once", result);
-    } catch (e: any) {
-      logger.err("err", `POST /run-once → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 500);
-    }
-  });
-
-  // POST /list-models — list models available through a runtime
-  //
-  // Body: { runtime: string, config?: ListModelsConfig }
-  // Returns: { models, source, fetchedAt, error? }
-  //
-  // The runtime parameter selects which AgentProvider answers (claude-code,
-  // codex, opencode, omlx). The config payload is provider-specific —
-  // claude-code/omlx honors `baseUrl` + `apiKey`, opencode honors `cliPath`.
-  // POST (not GET) because config may carry an apiKey we don't want logged
-  // in URLs or proxy access logs.
-  app.post("/list-models", async (c) => {
-    const body = (await c.req.json().catch(() => ({}))) as {
-      runtime?: string;
-      config?: import("../../core/providers/types.js").ListModelsConfig;
-    };
-    const runtime = body.runtime ?? "claude-code";
-    let provider;
-    try {
-      provider = getProvider(runtime);
-    } catch (e: any) {
-      return c.json({ status: "error", message: e.message }, 400);
-    }
-    if (!provider.listModels) {
-      return httpJson(c, "agent.list-models", {
-        models: [],
-        source: "static",
-        fetchedAt: Date.now(),
-        error: `runtime "${runtime}" does not support model listing`,
-      });
-    }
-    try {
-      const result = await provider.listModels(body.config);
-      return httpJson(c, "agent.list-models", result);
-    } catch (e: any) {
-      logger.err("err", `POST /list-models[${runtime}] → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 500);
-    }
-  });
-
-  // POST /completion — lightweight one-shot LLM call (no session)
-  app.post("/completion", async (c) => {
-    const body = (await c.req.json().catch(() => ({}))) as CompletionOptions;
-    if (!body.prompt) {
-      return c.json({ status: "error", message: "prompt is required" }, 400);
-    }
-    try {
-      const result = await completion(body);
-      return httpJson(c, "agent.completion", result);
-    } catch (e: any) {
-      logger.err("err", `POST /completion → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 500);
-    }
-  });
-
-  // ── Agent lifecycle (session-scoped) ──────────────────────────
-
-  // POST /start — start agent in a session (idempotent)
-  app.post("/start", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as {
-      provider?: string;
-      modelProvider?: string;
-      prompt?: string;
-      model?: string;
-      permissionMode?: string;
-      configDir?: string;
-      force?: boolean;
-      meta?: Record<string, unknown>;
-      extraArgs?: string[];
-      providerOptions?: Record<string, unknown>;
-      systemPrompt?: string;
-      appendSystemPrompt?: string;
-      allowedTools?: string[];
-      disallowedTools?: string[];
-      mcpServers?: Record<string, unknown>;
-      cwd?: string;
-      history?: import("../../history/types.js").CanonicalBlock[];
-      env?: Record<string, string>;
-    };
-
-    // Auto-create session if it doesn't exist (backward compat for "default")
-    const session = sessionManager.getOrCreateSession(sessionId, {
-      cwd: body.cwd,
-    });
-
-    // If agent is already alive and not forced, return existing session
-    if (session.process?.alive && !body.force) {
-      logger.log("route", `POST /start?session=${sessionId} → already_running`);
-      return httpJson(c, "agent.start", {
-        status: "already_running",
-        provider: getConfig().defaultProvider,
-        sessionId: session.process.sessionId ?? session.id,
-      });
-    }
-
-    // Kill existing
-    if (session.process?.alive) {
-      session.process.kill();
-    }
-
-    const provider = getProvider(body.provider ?? getConfig().defaultProvider);
-
-    // Persist initial prompt as user message
-    try {
-      const db = getDb();
-      db.prepare(`INSERT OR IGNORE INTO chat_sessions (id, label, type) VALUES (?, ?, 'main')`)
-        .run(sessionId, session.label ?? sessionId);
-      if (body.prompt) {
-        insertChatMessage(db, {
-          sessionId, actor: "user", kind: "text",
-          content: body.prompt,
-          meta: body.meta,
-        });
-      }
-    } catch { /* DB not ready — non-fatal */ }
-
-    const providerName = body.provider ?? getConfig().defaultProvider;
-    const model = body.model ?? getConfig().model;
-    const permissionMode = body.permissionMode;
-    const configDir = body.configDir;
-    const extraArgs = body.extraArgs;
-
-    try {
-      const proc = await spawnWithPool(provider, {
-        cwd: session.cwd,
-        prompt: body.prompt,
-        model,
-        permissionMode: permissionMode as any,
-        configDir,
-        env: { ...body.env, SNA_SESSION_ID: sessionId },
-        history: body.history,
-        extraArgs,
-        providerOptions: body.providerOptions,
-        systemPrompt: body.systemPrompt,
-        appendSystemPrompt: body.appendSystemPrompt,
-        allowedTools: body.allowedTools,
-        disallowedTools: body.disallowedTools,
-        mcpServers: body.mcpServers as any,
-      });
-
-      sessionManager.setProcess(sessionId, proc);
-      sessionManager.saveStartConfig(sessionId, { provider: providerName, modelProvider: body.modelProvider, model, cwd: session.cwd, permissionMode, configDir, extraArgs, providerOptions: body.providerOptions });
-      logger.log("route", `POST /start?session=${sessionId} → started`);
-
-      return httpJson(c, "agent.start", {
-        status: "started",
-        provider: provider.name,
-        sessionId: session.id,
-      });
-    } catch (e: any) {
-      logger.err("err", `POST /start?session=${sessionId} failed: ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 500);
-    }
-  });
-
-  // POST /send — send a message to the agent
-  app.post("/send", async (c) => {
-    const sessionId = getSessionId(c);
-    const session = sessionManager.getSession(sessionId);
-
-    if (!session?.process?.alive) {
-      logger.err("err", `POST /send?session=${sessionId} → no active session`);
-      return c.json(
-        { status: "error", message: `No active agent session "${sessionId}". Call POST /start first.` },
-        400,
-      );
-    }
-
-    const body = (await c.req.json().catch(() => ({}))) as {
-      message?: string;
-      images?: Array<{ base64: string; mimeType: string }>;
-      meta?: Record<string, unknown>;
-    };
-    if (!body.message && !body.images?.length) {
-      logger.err("err", `POST /send?session=${sessionId} → empty message`);
-      return c.json({ status: "error", message: "message or images required" }, 400);
-    }
-
-    // Persist user message. Attached images/files are stashed into the same
-    // row via embed refs (`![](embed://<id>)` in content) + the `embeds` JSON
-    // column holding the EmbedRecord dictionary.
-    const userText = body.message ?? "";
-    const meta: Record<string, unknown> = body.meta ? { ...body.meta } : {};
-    const embeds: Record<string, EmbedRecord> = {};
-    let contentText = userText;
-    if (body.images?.length) {
-      const saved = saveEmbeds(sessionId, body.images);
-      const refs = saved.map(({ id, record }) => {
-        embeds[id] = record;
-        return formatEmbedRef(id);
-      });
-      contentText = userText ? `${userText}\n${refs.join(" ")}` : refs.join(" ");
-    }
-    try {
-      const db = getDb();
-      db.prepare(`INSERT OR IGNORE INTO chat_sessions (id, label, type) VALUES (?, ?, 'main')`)
-        .run(sessionId, session.label ?? sessionId);
-      insertChatMessage(db, {
-        sessionId,
-        actor: "user",
-        kind: "text",
-        content: contentText,
-        embeds: Object.keys(embeds).length > 0 ? embeds : undefined,
-        meta: Object.keys(meta).length > 0 ? meta : undefined,
-      });
-    } catch { /* DB write failure is non-fatal */ }
-
-    // Broadcast user message to agent.subscribe listeners (multi-client sync)
-    sessionManager.pushEvent(sessionId, {
-      type: "user_message",
-      message: contentText,
-      data: Object.keys(meta).length > 0 ? meta : undefined,
-      timestamp: Date.now(),
-    });
-
-    sessionManager.updateSessionState(sessionId, "processing");
-    sessionManager.touch(sessionId);
-
-    // Build content: plain string or content block array with images
-    if (body.images?.length) {
-      const content: import("../../core/providers/types.js").ContentBlock[] = [
-        ...body.images.map((img) => ({
-          type: "image" as const,
-          source: { type: "base64" as const, media_type: img.mimeType, data: img.base64 },
-        })),
-        ...(body.message ? [{ type: "text" as const, text: body.message }] : []),
-      ];
-      logger.log("route", `POST /send?session=${sessionId} → ${body.images.length} image(s) + "${(body.message ?? "").slice(0, 40)}"`);
-      session.process.send(content);
-    } else {
-      logger.log("route", `POST /send?session=${sessionId} → "${body.message!.slice(0, 80)}"`);
-      session.process.send(body.message!);
-    }
-    return httpJson(c, "agent.send", { status: "sent" });
-  });
-
-  // GET /events — SSE stream (stays open indefinitely)
-  app.get("/events", (c) => {
-    const sessionId = getSessionId(c);
-    const session = sessionManager.getOrCreateSession(sessionId);
-
-    const sinceParam = c.req.query("since");
-    const sinceCursor = sinceParam ? parseInt(sinceParam, 10) : session.eventCounter;
-
-    return streamSSE(c, async (stream) => {
-      const KEEPALIVE_MS = getConfig().keepaliveIntervalMs;
-      const signal = c.req.raw.signal;
-
-      // Queue bridges sync event callbacks → async SSE writes
-      const queue: Array<{ cursor: number; event: AgentEvent }> = [];
-      let wakeUp: (() => void) | null = null;
-
-      const unsub = sessionManager.onSessionEvent(sessionId, (eventCursor, event) => {
-        queue.push({ cursor: eventCursor, event });
-        const fn = wakeUp; wakeUp = null; fn?.();
-      });
-      signal.addEventListener("abort", () => { const fn = wakeUp; wakeUp = null; fn?.(); });
-
-      try {
-        // Replay buffer history for requested cursor range
-        let cursor = sinceCursor;
-        if (cursor < session.eventCounter) {
-          const startIdx = Math.max(
-            0,
-            session.eventBuffer.length - (session.eventCounter - cursor),
-          );
-          for (const event of session.eventBuffer.slice(startIdx)) {
-            cursor++;
-            await stream.writeSSE({ id: String(cursor), data: JSON.stringify(event) });
-          }
-        } else {
-          cursor = session.eventCounter;
-        }
-
-        // Drop real-time events that were already covered by buffer replay.
-        // cursor=-1 events (transient deltas) always pass through.
-        while (queue.length > 0 && queue[0].cursor !== -1 && queue[0].cursor <= cursor) queue.shift();
-
-        // Event-driven loop — no polling
-        while (!signal.aborted) {
-          if (queue.length === 0) {
-            await Promise.race([
-              new Promise<void>((r) => { wakeUp = r; }),
-              new Promise<void>((r) => setTimeout(r, KEEPALIVE_MS)),
-            ]);
-          }
-
-          if (signal.aborted) break;
-
-          if (queue.length > 0) {
-            while (queue.length > 0) {
-              const item = queue.shift()!;
-              // cursor=-1 = transient event (assistant_delta) — send without SSE id
-              if (item.cursor === -1) {
-                await stream.writeSSE({ data: JSON.stringify(item.event) });
-              } else {
-                await stream.writeSSE({ id: String(item.cursor), data: JSON.stringify(item.event) });
-              }
-            }
-          } else {
-            // Keepalive — no events arrived within KEEPALIVE_MS
-            await stream.writeSSE({ data: "" });
-          }
-        }
-      } finally {
-        unsub();
-      }
-    });
-  });
-
-  // POST /restart — kill + re-spawn with merged config
-  // Same provider: --resume (native session continuity)
-  // Different provider: DB history injection (cross-provider handoff)
-  app.post("/restart", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as {
-      provider?: string;
-      modelProvider?: string;
-      model?: string;
-      cwd?: string;
-      permissionMode?: string;
-      configDir?: string;
-      extraArgs?: string[];
-      providerOptions?: Record<string, unknown>;
-      systemPrompt?: string;
-      appendSystemPrompt?: string;
-      allowedTools?: string[];
-      disallowedTools?: string[];
-      mcpServers?: Record<string, unknown>;
-      env?: Record<string, string>;
-    };
-
-    try {
-      const session = sessionManager.getSession(sessionId);
-      if (!session) return c.json({ status: "error", message: "Session not found" }, 404);
-
-      const prevProvider = session.config?.provider;
-      const nextProvider = body.provider ?? prevProvider;
-      if (!nextProvider) return c.json({ status: "error", message: "Provider is required" }, 400);
-      const nextProv = getProvider(nextProvider);
-
-      // Typed SpawnOptions fields are forwarded from the request body directly
-      // — they are not part of StartConfig persistence because they describe
-      // the current spawn, not the session's identity.
-      const typedOpts = {
-        systemPrompt: body.systemPrompt,
-        appendSystemPrompt: body.appendSystemPrompt,
-        allowedTools: body.allowedTools,
-        disallowedTools: body.disallowedTools,
-        mcpServers: body.mcpServers as any,
-      };
-
-      // Single unified path. restartSession handles kill + persist; spawnWithPool
-      // hides the pooled/non-pooled branching so this site can't forget the
-      // runtime pool for pooled providers.
-      const { config } = await sessionManager.restartSession(sessionId, body, async (cfg) => {
-        const prov = getProvider(cfg.provider);
-        const providerChanged = prevProvider && cfg.provider !== prevProvider;
-
-        if (providerChanged) {
-          // Cross-provider: inject DB history
-          const history = buildCanonicalFromDb(sessionId);
-          logger.log("route", `restart: provider changed ${prevProvider} → ${cfg.provider}, using DB history (${history.length} msgs)`);
-          return spawnWithPool(prov, {
-            cwd: session.cwd,
-            model: cfg.model,
-            permissionMode: cfg.permissionMode as any,
-            configDir: cfg.configDir,
-            env: { ...body.env, SNA_SESSION_ID: sessionId },
-            history: history.length > 0 ? history : undefined,
-            extraArgs: cfg.extraArgs,
-            providerOptions: cfg.providerOptions,
-            ...typedOpts,
-          });
-        }
-
-        // Same provider: native resume via resumeSessionId
-        return spawnWithPool(prov, {
-          cwd: session.cwd,
-          model: cfg.model,
-          permissionMode: cfg.permissionMode as any,
-          configDir: cfg.configDir,
-          env: { ...body.env, SNA_SESSION_ID: sessionId },
-          resumeSessionId: session.ccSessionId ?? undefined,
-          extraArgs: cfg.extraArgs,
-          providerOptions: cfg.providerOptions,
-          ...typedOpts,
-        });
-      });
-
-      logger.log("route", `POST /restart?session=${sessionId} → restarted (${config.provider})`);
-      return httpJson(c, "agent.restart", {
-        status: "restarted",
-        provider: config.provider,
-        sessionId,
-      });
-    } catch (e: any) {
-      logger.err("err", `POST /restart?session=${sessionId} → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 500);
-    }
-  });
-
-  // POST /resume — resume session with DB history auto-injected
-  app.post("/resume", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as {
-      prompt?: string;
-      model?: string;
-      permissionMode?: string;
-      configDir?: string;
-      provider?: string;
-      modelProvider?: string;
-      extraArgs?: string[];
-      providerOptions?: Record<string, unknown>;
-      systemPrompt?: string;
-      appendSystemPrompt?: string;
-      allowedTools?: string[];
-      disallowedTools?: string[];
-      mcpServers?: Record<string, unknown>;
-      env?: Record<string, string>;
-    };
-
-    const session = sessionManager.getOrCreateSession(sessionId);
-    if (session.process?.alive) {
-      return c.json({ status: "error", message: "Session already running. Use agent.send instead." }, 400);
-    }
-
-    const history = buildCanonicalFromDb(sessionId);
-    if (history.length === 0 && !body.prompt) {
-      return c.json({ status: "error", message: "No history in DB — nothing to resume." }, 400);
-    }
-
-    const providerName = body.provider ?? getConfig().defaultProvider;
-    const providerChanged = session.config && session.config.provider !== providerName;
-    const model = body.model ?? session.config?.model ?? getConfig().model;
-    const permissionMode = body.permissionMode ?? session.config?.permissionMode;
-    const configDir = providerChanged ? body.configDir : (body.configDir ?? session.config?.configDir);
-    // Provider-specific fields: inherit only when provider is unchanged.
-    const extraArgs = providerChanged ? body.extraArgs : (body.extraArgs ?? session.config?.extraArgs);
-    const providerOptions = providerChanged ? body.providerOptions : (body.providerOptions ?? session.config?.providerOptions);
-    // modelProvider: inherit only when provider is unchanged AND caller didn't
-    // supply one; otherwise prefer the caller's explicit value.
-    const modelProvider = body.modelProvider ?? (providerChanged ? undefined : session.config?.modelProvider);
-    const provider = getProvider(providerName);
-
-    try {
-      const proc = await spawnWithPool(provider, {
-        cwd: session.cwd,
-        prompt: body.prompt,
-        model,
-        permissionMode: permissionMode as any,
-        configDir,
-        env: { ...body.env, SNA_SESSION_ID: sessionId },
-        history: history.length > 0 ? history : undefined,
-        extraArgs,
-        providerOptions,
-        systemPrompt: body.systemPrompt,
-        appendSystemPrompt: body.appendSystemPrompt,
-        allowedTools: body.allowedTools,
-        disallowedTools: body.disallowedTools,
-        mcpServers: body.mcpServers as any,
-      });
-      sessionManager.setProcess(sessionId, proc, "resumed");
-      sessionManager.saveStartConfig(sessionId, { provider: providerName, modelProvider, model, cwd: session.cwd, permissionMode, configDir, extraArgs, providerOptions });
-      logger.log("route", `POST /resume?session=${sessionId} → resumed (${history.length} history msgs)`);
-      return httpJson(c, "agent.resume", {
-        status: "resumed",
-        provider: providerName,
-        sessionId: session.id,
-        historyCount: history.length,
-      });
-    } catch (e: any) {
-      logger.err("err", `POST /resume?session=${sessionId} → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 500);
-    }
-  });
-
-  // POST /interrupt — interrupt current turn, process stays alive
-  app.post("/interrupt", async (c) => {
-    const sessionId = getSessionId(c);
-    const interrupted = sessionManager.interruptSession(sessionId);
-    return httpJson(c, "agent.interrupt", { status: interrupted ? "interrupted" : "no_session" });
-  });
-
-  // POST /set-model — change model at runtime, no restart
-  app.post("/set-model", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as { model?: string };
-    if (!body.model) return c.json({ status: "error", message: "model is required" }, 400);
-    const updated = sessionManager.setSessionModel(sessionId, body.model);
-    return httpJson(c, "agent.set-model", { status: updated ? "updated" : "no_session", model: body.model });
-  });
-
-  // POST /set-permission-mode — change permission mode at runtime, no restart
-  app.post("/set-permission-mode", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as { permissionMode?: string };
-    if (!body.permissionMode) return c.json({ status: "error", message: "permissionMode is required" }, 400);
-    const updated = sessionManager.setSessionPermissionMode(sessionId, body.permissionMode);
-    return httpJson(c, "agent.set-permission-mode", { status: updated ? "updated" : "no_session", permissionMode: body.permissionMode });
-  });
-
-  // PATCH /session — unified PATCH mutator. Applies the patch in-place where
-  // the provider supports it (codex per-turn cwd/model/sandbox override,
-  // claude-code control_request for model/permissionMode); leftover fields
-  // (claude-code cwd, cross-runtime changes) drive a respawn-with-history.
-  // Consumers don't have to switch on runtime — the response reports which
-  // path was taken via `applied`. See #21.
-  app.patch("/session", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as {
-      cwd?: string;
-      model?: string;
-      permissionMode?: string;
-    };
-    // SessionPatch fields are the only thing PATCH /session accepts today.
-    // Unknown fields are silently ignored — consumers should use /restart
-    // when they need to change provider / configDir / providerOptions etc.
-    const patch = {
-      ...(body.cwd !== undefined ? { cwd: body.cwd } : {}),
-      ...(body.model !== undefined ? { model: body.model } : {}),
-      ...(body.permissionMode !== undefined ? { permissionMode: body.permissionMode } : {}),
-    };
-
-    const session = sessionManager.getSession(sessionId);
-    if (!session) return c.json({ status: "error", message: "Session not found" }, 404);
-
-    try {
-      const result = await sessionManager.applySessionPatch(sessionId, patch, async (cfg) => {
-        // Respawn path: spawn a fresh process with the merged config and
-        // replay DB history so the agent picks up where the prior runtime
-        // left off. Same recipe restartSession uses for cross-provider hops.
-        const prov = getProvider(cfg.provider);
-        const history = buildCanonicalFromDb(sessionId);
-        return spawnWithPool(prov, {
-          cwd: cfg.cwd,
-          model: cfg.model,
-          permissionMode: cfg.permissionMode as any,
-          configDir: cfg.configDir,
-          env: { SNA_SESSION_ID: sessionId },
-          history: history.length > 0 ? history : undefined,
-          extraArgs: cfg.extraArgs,
-          providerOptions: cfg.providerOptions,
-        });
-      });
-      logger.log("route", `PATCH /session?session=${sessionId} → ${result.applied} fields=[${result.fields.join(",")}] rt=${result.runtimeId}`);
-      return httpJson(c, "agent.session.patch", {
-        status: "updated",
-        applied: result.applied,
-        runtimeId: result.runtimeId,
-        fields: result.fields,
-      });
-    } catch (e: any) {
-      logger.err("err", `PATCH /session?session=${sessionId} → ${e.message}`);
-      return c.json({ status: "error", message: e.message }, 400);
-    }
-  });
-
-  // POST /kill
-  app.post("/kill", async (c) => {
-    const sessionId = getSessionId(c);
-    const session = sessionManager.getSession(sessionId);
-
-    if (!session?.process?.alive) {
-      return httpJson(c, "agent.kill", { status: "no_session" });
-    }
-
-    // For pooled providers (Codex, OpenCode), close only the thread.
-    // For non-pooled providers (Claude Code), kill the entire process.
-    const providerName = session.config?.provider;
-    if (providerName) {
-      const provider = getProvider(providerName);
-      if (provider.supportsRuntimePooling) {
-        session.process.closeThread();
-      } else {
-        session.process.kill();
-      }
-    } else {
-      // No provider info — fall back to kill (safe default)
-      session.process.kill();
-    }
-
-    return httpJson(c, "agent.kill", { status: "killed" });
-  });
-
-  // GET /status
-  app.get("/status", (c) => {
-    const sessionId = getSessionId(c);
-    const session = sessionManager.getSession(sessionId);
-    const alive = session?.process?.alive ?? false;
-    let messageCount = 0;
-    let lastMessage: { actor: string; kind: string; content: string; created_at: string } | null = null;
-    try {
-      const db = getDb();
-      const count = db.prepare("SELECT COUNT(*) as c FROM chat_messages WHERE session_id = ?").get(sessionId) as any;
-      messageCount = count?.c ?? 0;
-      const last = db.prepare("SELECT actor, kind, content, created_at FROM chat_messages WHERE session_id = ? ORDER BY id DESC LIMIT 1").get(sessionId) as any;
-      if (last) lastMessage = { actor: last.actor, kind: last.kind, content: last.content, created_at: last.created_at };
-    } catch {}
-    return httpJson(c, "agent.status", {
-      alive,
-      agentStatus: !alive ? "disconnected" : (session?.state === "processing" ? "busy" : "idle"),
-      sessionId: session?.process?.sessionId ?? null,
-      ccSessionId: session?.ccSessionId ?? null,
-      eventCount: session?.eventCounter ?? 0,
-      messageCount,
-      lastMessage,
-      config: session?.config ?? null,
-    });
-  });
-
-  // ── Permission approval flow ────────────────────────────────────
-
-  // POST /permission-request — called by hook.ts (sync) to submit a request and wait
-  app.post("/permission-request", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as {
-      tool_name?: string;
-      tool_input?: Record<string, unknown>;
-    };
-
-    logger.log("route", `POST /permission-request?session=${sessionId} → ${body.tool_name}`);
-
-    const result = await sessionManager.createPendingPermission(sessionId, body);
-    return c.json({ approved: result });
-  });
-
-  // POST /permission-respond — called by UI to approve/deny
-  app.post("/permission-respond", async (c) => {
-    const sessionId = getSessionId(c);
-    const body = (await c.req.json().catch(() => ({}))) as { approved?: boolean };
-    const approved = body.approved ?? false;
-
-    const resolved = sessionManager.resolvePendingPermission(sessionId, approved);
-    if (!resolved) {
-      return c.json({ status: "error", message: "No pending permission request" }, 404);
-    }
-
-    logger.log("route", `POST /permission-respond?session=${sessionId} → ${approved ? "approved" : "denied"}`);
-    return httpJson(c, "permission.respond", { status: approved ? "approved" : "denied" });
-  });
-
-  // GET /permission-pending — UI polls this to check for pending requests
-  // Always returns { pending: Array } for consistent typing
-  app.get("/permission-pending", (c) => {
-    const sessionId = c.req.query("session");
-
-    if (sessionId) {
-      const pending = sessionManager.getPendingPermission(sessionId);
-      return httpJson(c, "permission.pending", { pending: pending ? [{ sessionId, ...pending }] : [] });
-    }
-
-    // No session specified — return all pending
-    return httpJson(c, "permission.pending", { pending: sessionManager.getAllPendingPermissions() });
-  });
-
-  return app;
-}
+export { runOnce } from "../run-once.js";
+export type { RunOnceOptions, RunOnceResult } from "../run-once.js";
